# swagger_config.yaml
swagger: "2.0"
info:
  title: "Volumx Business API"
  description: "Endpoints for managing businesses and user rights"
  version: "1.0"
schemes:
  - http
  - https
...
paths:
...
  /api/v1/auth/register:
    post:
      tags:
        - Authentication
      summary: User registration
      description: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
                first_name:
                  type: string
                last_name:
                  type: string
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Incomplete registration data
        '500':
          description: Internal server error
  /api/v1/auth/login:
    post:
      tags:
        - Authentication
      summary: User login and token generation
      description: Authenticate and generate access tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Logged in successfully
          content:
            application/json:
        '400':
          description: Email and password are required
        '401':
          description: Invalid email or password
...